# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoRa KT API",
    "version": "1.0.0",
    "description": "REST API for team management, checkpoints, RFID access control, LoRa telemetry, and map data."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Relative to application root"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "webhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Optional server-level secret for the ingest endpoint. When LORA_WEBHOOK_SECRET is set (and not the default placeholder), every POST /api/ingest must include this header."
      }
    },
    "schemas": {
...
      "GroupCheckpoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "position"
        ]
      },
      "CheckpointGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "checkpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupCheckpoint"
            }
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/groups": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "List checkpoint groups",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "groups": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CheckpointGroup"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Create group",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "description": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "checkpoint_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created group"
          }
        }
      }
    },
...
    "/api/score-rules": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "List score rules",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "checkpoint_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "group_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rules"
          }
        }
      },
      "post": {
        "tags": [
          "scores"
        ],
        "summary": "Create or update a score rule",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "checkpoint_id",
                  "group_id",
                  "rules"
                ],
                "properties": {
                  "checkpoint_id": {
                    "type": "integer"
                  },
                  "group_id": {
                    "type": "integer"
                  },
                  "rules": {
                    "type": "object",
                    "description": "Rule object with optional field_rules and time_race. field_rules maps field keys to a rule or list of rules: mapping {type:'mapping', map:{input:output}}, interpolate {type:'interpolate', points:[[x,y],...]}, multiplier {type:'multiplier', factor}, found {type:'found', checkpoint_ids:[int], points_per:number}, deviation {type:'deviation', target, max_points, penalty_points, penalty_distance, min_points}. time_race uses checkin times between checkpoints: {start_checkpoint_id, end_checkpoint_id, min_points, max_points}.",
                    "example": {
                      "field_rules": {
                        "Distance": [
                          {
                            "type": "deviation",
                            "target": 0,
                            "max_points": 100,
                            "penalty_points": 1,
                            "penalty_distance": 1,
                            "min_points": 0
                          }
                        ],
                        "Found": {
                          "type": "found",
                          "checkpoint_ids": [
                            12,
                            14
                          ],
                          "points_per": 5
                        }
                      },
                      "total_fields": [
                        "Distance",
                        "Found"
                      ],
                      "time_race": {
                        "start_checkpoint_id": 3,
                        "end_checkpoint_id": 7,
                        "min_points": 10,
                        "max_points": 100
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated rule"
          },
          "201": {
            "description": "Created rule"
          }
        }
      }
    },
    "/api/score-rules/fields": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Get score rule fields",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "checkpoint_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "group_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available scoring fields"
          }
        }
      }
    },
    "/api/score-rules/{rule_id}": {
      "parameters": [
        {
          "name": "rule_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "delete": {
        "tags": [
          "scores"
        ],
        "summary": "Delete a score rule",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    }
  }
}